+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - `repka.repositories.base.AsyncBaseRepo.update_or_insert_first` - same as `update_or_insert_many_by_field` but for one
   entity (#58 by @kosyan62)
 
+### Changed
+
+- py38 win32 support (#62 by @potykion)
+
 ## 3.1.1 - 2020-10-30
 
 ### Changed